Updated to support early CoExistence notification to the WiFi device
ENABLE_EARLY_NOTIFICATION - If enabled, the firmware will indicate that
the RF event is active 900us before the actual ADV or Connection start
event. Note that this will only work if the BLESS is configured in
DeepSleep mode after each ADV or Connection event.

Updated project to PSoC Creator 3.3 SP1